pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]      # PEP 508 specifications.
build-backend = "setuptools.build_meta"

[project]
name = "axon-synapse-demo-tictactoe"
version = "0.0.1"
requires-python = ">=3.10"
description = "Axon Synapse Demo - Tic Tac Toe"
dependencies = ["axon-python-synapse", "aiohttp_jinja2", "click"]
authors = [{ name = "Ferhat Ayaz", email = "ferhat.ayaz@holisticon.de" }]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://www.holisticon.de"
# Documentation = "https://readthedocs.org"
# Repository = "https://github.com/me/spam.git"
# Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"
